refactor(protectedroute): refactor to ts and use children
